Refactored Events cog into subclasses Much easier to develop Events now! Just make the subclass and add to the EventFactory. All of this is managed by the EventManager cog.
